#17 [REFACTOR] TrendingCarousel을 TrendingHero로 분리 및 서버 컴포넌트 구조 개선
